% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_compiler.R
\name{activity_compiler}
\alias{activity_compiler}
\title{convert a single activity list into a dataframe}
\usage{
activity_compiler(x, columns)
}
\arguments{
\item{x}{a list containing details of a single Strava activity}

\item{columns}{a character vector of all the columns in the list of Strava activities. Produced automatically in \code{\link{get_activities_dataframe}}. Leave blank if running for a single activity list.}
}
\value{
dataframe where every column is an item from a list. Any missing columns rom the total number of columns
}
\description{
convert a single activity list into a dataframe
}
\details{
used internally in \code{\link{get_activities_dataframe}}
}
\author{
Daniel Padfield
}
\concept{
posttoken
}

