% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlind_fun.R
\name{athlind_fun}
\alias{athlind_fun}
\title{Get data for a single athlete}
\usage{
athlind_fun(athl_num)
}
\arguments{
\item{athl_num}{numeric athlete id used by Strava}
}
\value{
A list with elements for the athlete id, units of measurement, location, monthly data, year-to-date data, and an all-time summary.
}
\description{
Get data for a single athlete by web scraping, does not require authentication.
}
\concept{
notoken
}
