% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_activities.R
\name{compile_activities}
\alias{compile_activities}
\alias{compile_activities.actlist}
\title{converts a list of activities into a dataframe}
\usage{
compile_activities(actlist, ...)

\method{compile_activities}{actlist}(actlist, acts = NULL, ...)
}
\arguments{
\item{actlist}{an \code{actlist} object returned by \code{\link{compile_activities}}}

\item{...}{arguments passed to or from other methods}

\item{acts}{numeric indicating which activities to compile starting with most recent, defaults to all}
}
\value{
dataframe where each row is a different activity
}
\description{
converts a list of activities into a dataframe
}
\details{
each activity has a value for every column present across all activities, with NAs populating empty values
}
\examples{
 
\dontrun{
stoken <- httr::config(ttoken = strava_oauth(app_name, app_client_id, app_secret, cache = TRUE))

acts <- get_activity_list(stoken)

acts_data <- compile_activities(acts)
}
}
\author{
Daniel Padfield
}
\concept{
token
}

