% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_segment.R
\name{compile_segment}
\alias{compile_segment}
\title{Compile information on a segment}
\usage{
compile_segment(segment_id, stoken)
}
\arguments{
\item{segment_id}{a Strava segment id}

\item{stoken}{A \code{\link[httr]{config}} object created using the \code{\link{strava_oauth}} function}
}
\value{
dataframe of all information given in a call from \code{\link{get_segment}}
}
\description{
Compile generation information on a segment
}
\details{
compiles information a segment by calling \code{\link{get_segment}} internally
}
\concept{token}
