% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_Polyline.R
\name{decode_Polyline}
\alias{decode_Polyline}
\title{Decode Google polyline of an activity into latitude and longitude}
\usage{
decode_Polyline(Google_polyline)
}
\arguments{
\item{Google_polyline}{character element of Google polyline of an activity}
}
\value{
A vector of latitudes an longitudes in decimals separated by a comma
}
\description{
Decodes Google's polyline that are given in get_activity() and get_activity_list()
}
\details{
When getting an activity using get_activity() a Google polyline is returned as one of the outputs. This function converts the polyline into latitude and longitude coordinates suitable for plotting. Function is used internally within \code{\link{get_LatLon}} and \code{\link{get_all_LatLon}}.
}
\examples{
\dontrun{
decode_Polyline()}
}
\author{
Daniel Padfield
}
\references{
\link{https://s4rdd.blogspot.co.uk/2012/12/google-maps-api-decoding-polylines-for.html?showComment=1473004506791#c3610119369153401460}
}
\concept{
notoken
}

