% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{calculates cumulative distance from activity longitude latitude points}
\usage{
distance(data, lon, lat)
}
\arguments{
\item{data}{the data frame containing the latitude and longitude points}

\item{lon}{the column name for your longitude points}

\item{lat}{the column name for your latitude points}
}
\value{
vector of cumulative distances
}
\description{
calculates cumulative distance from activity longitude latitude points
}
\author{
Daniel Padfield
}
\concept{
token
}

