% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow_fun.R
\name{follow_fun}
\alias{follow_fun}
\title{Get athlete follow data}
\usage{
follow_fun(prsd)
}
\arguments{
\item{prsd}{parsed input list}
}
\value{
A data frame of counts of followers and following for the athlete. An empty list is returned if none found.
}
\description{
Get athlete follow data, used internally in \code{\link{athl_fun}}
}
\concept{notoken}
