% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_LatLon.R
\name{get_all_LatLon}
\alias{get_all_LatLon}
\title{get latitude and longitude for multiple activities}
\usage{
get_all_LatLon(id_col, parent_data)
}
\arguments{
\item{id_col}{the column that you want to be used as an identifier for the dataframe of latitude and longitude coordinates}

\item{parent_data}{the dataframe that contains the Strava activity data}
}
\value{
dataframe of latitude and longitudes with a column for each unique identifier
}
\description{
get latitude and longitude for multiple activities
}
\details{
uses \code{\link{get_LatLon}} and \code{\link{decode_Polyline}} to produce a dataframe of latitudes and longitudes
}
\examples{
\dontrun{
get_all_LatLon('upload_id', acts_data)
}
}
\author{
Daniel Padfield
}
\concept{
token
}
