% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elev_prof.R
\name{get_elev_prof}
\alias{get_elev_prof}
\title{Create elevation profiles from activitiy data}
\usage{
get_elev_prof(act_data, acts = 1, key)
}
\arguments{
\item{act_data}{a list of Strava activities derived from \code{\link{get_activity_list}}}

\item{acts}{numeric value indicating which elements of \code{act_data} to plot}

\item{key}{chr string of Google API key for elevation data, passed to \code{\link[rgbif]{elevation}}, see details}
}
\value{
A \code{ggplot} of elevation profiles, facetted by activity id, date
}
\description{
Create elevation profiles from activitiy data
}
\details{
The Google API key is easy to obtain, follow instructions here: https://developers.google.com/maps/documentation/elevation/#api_key
}
\examples{
\dontrun{
# get my activities
stoken <- httr::config(ttoken = strava_oauth(app_name, app_client_id, app_secret, cache = TRUE))
my_acts <- get_activity_list(stoken)

# your unique key
mykey <- 'Get Google API key'
get_elev_prof(my_acts, acts = 1:2, key = mykey)
}
}
\author{
Daniel Padfield
}
\concept{
token
}

