% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_following.R
\name{get_following}
\alias{get_following}
\title{Get followers, friends, or both-following}
\usage{
get_following(following, stoken, id = NULL)
}
\arguments{
\item{following}{string equal to ``friends", ``followers", or ``both-following"}

\item{stoken}{A \code{\link[httr]{config}} object created using the \code{\link{strava_oauth}} function}

\item{id}{string or integer of athlete, taken from \code{stoken} if \code{NULL}}
}
\value{
Data from an API request.
}
\description{
Get followers or friends of the athlete or both-following relative to another user
}
\details{
Requires authentication stoken using the \code{\link{strava_oauth}} function and a user-created API on the strava website.
}
\examples{
\dontrun{
# create authentication token
# requires user created app name, id, and secret from Strava website
stoken <- httr::config(token = strava_oauth(app_name, app_client_id,
	app_secret, cache = TRUE))

get_following('friends', stoken)
}
}
\concept{
token
}

