% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heat_map.R
\name{get_heat_map}
\alias{get_heat_map}
\title{Makes a heat map from your activity data}
\usage{
get_heat_map(act_data, alpha = NULL, f = 1)
}
\arguments{
\item{act_data}{a dataframe of Strava activities derived from \code{\link{compile_activities}}}

\item{alpha}{the opacity of the line desired. A single activity should be 1. Defaults to 0.5}

\item{f}{number specifying the fraction by which the range should be extended for the bounding box of the activities, passed to \code{\link[ggmap]{make_bbox}}}
}
\value{
plot of activity on a Google map
}
\description{
Makes a heat map from your activity data
}
\details{
uses \code{\link{get_all_LatLon}} to produce a dataframe of latitudes and longitudes to use in the map. Uses {ggmap} to produce map and ggplot2 it
}
\examples{
\dontrun{
stoken <- httr::config(ttoken = strava_oauth(app_name, app_client_id, app_secret, cache = TRUE))

acts <- get_activity_list(stoken, 2837007)

acts_data <- compile_activities(acts)

get_heat_map(acts_data[1,], alpha = 1)
}
}
\author{
Daniel Padfield
}
\concept{
token
}

