% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loc_fun.R
\name{loc_fun}
\alias{loc_fun}
\title{Get athlete location}
\usage{
loc_fun(xml_in)
}
\arguments{
\item{xml_in}{input HTMLInternalDocument from \code{\link[RCurl]{getURL}}}
}
\description{
Get location of an athlete, used internally in \code{\link{athl_fun}}
}
\concept{
notoken
}

