% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_fun.R
\name{location_fun}
\alias{location_fun}
\title{Get athlete location}
\usage{
location_fun(prsd)
}
\arguments{
\item{prsd}{parsed input list}
}
\value{
A character string of the athlete location
}
\description{
Get athlete location, used internally in \code{\link{athl_fun}}
}
\concept{notoken}
