% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_fun.R
\name{monthly_fun}
\alias{monthly_fun}
\title{Get distance and time for current month}
\usage{
monthly_fun(prsd)
}
\arguments{
\item{prsd}{parsed input list}
}
\value{
A data frame of the current monthly distance and time for the athlete. An empty list is returned if none found.
}
\description{
Get distance and time for current month, used internally in \code{\link{athl_fun}}
}
\concept{notoken}
