% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratelimit.R
\name{ratelimit}
\alias{ratelimit}
\title{Generate the ratelimit indicator}
\usage{
ratelimit(req)
}
\arguments{
\item{req}{value returned from the \code{\link[httr]{GET}} function, used internally in other functions}
}
\value{
A variable for the current limits.
}
\description{
Checks the ratelimit values after the last request and stores the left requests in a global variable.
}
\details{
Requests to the Strava API are rate-limited. The default rate limit allows 600 requests every 15 minutes, with up to 30,000 requests per day.  See the documentation at \url{https://strava.github.io/api/#access}.
}
\concept{token}
