% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recent_fun.R
\name{recent_fun}
\alias{recent_fun}
\title{Get last three recent activities}
\usage{
recent_fun(prsd)
}
\arguments{
\item{prsd}{parsed input list}
}
\value{
A data frame of recent activities for the athlete.  An empty list is returned if none found.
}
\description{
Get last three recent activities, used internally in \code{\link{athl_fun}}
}
\concept{notoken}
