% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seltime_fun.R
\name{seltime_fun}
\alias{seltime_fun}
\title{Format before and after arguments for API query}
\usage{
seltime_fun(dtin, before = TRUE)
}
\arguments{
\item{dtin}{Date object for \code{before} or \code{after} inputs}

\item{before}{logical indicattng if input is \code{before}}
}
\value{
A numeric object as an epoch timestamp
}
\description{
Format before and after arguments for API query
}
\examples{
# convert to epoch timestamp
seltime_fun(Sys.Date())

# back to original 
as.POSIXct(seltime_fun(Sys.Date(), before = FALSE), tz = Sys.timezone(), origin = '1970-01-01')
}
