% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summ_fun.R
\name{summ_fun}
\alias{summ_fun}
\title{Get summary data for an athlete}
\usage{
summ_fun(prsd)
}
\arguments{
\item{prsd}{parsed \code{\link[XML]{htmlTreeParse}} list}
}
\value{
two-element list
}
\description{
Get summary data for an athlete, year to date and total units of measurement. Used internally in \code{\link{athl_fun}}.
}
\concept{
notoken
}

