% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophy_fun.R
\name{trophy_fun}
\alias{trophy_fun}
\title{Get athlete trophies}
\usage{
trophy_fun(prsd)
}
\arguments{
\item{prsd}{parsed input list}
}
\value{
A data frame of trophies for the athlete. An empty list is returned if none found.
}
\description{
Get athlete trophies, used internally in \code{\link{athl_fun}}
}
\concept{notoken}
