% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_fun.R
\name{units_fun}
\alias{units_fun}
\title{Get units of measurement}
\usage{
units_fun(prsd)
}
\arguments{
\item{prsd}{parsed input list}
}
\description{
Get units of measurement, used internally in \code{\link{athl_fun}}
}
\concept{notoken}
