% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_activities.R
\name{url_activities}
\alias{url_activities}
\title{Set the url of activities for different activity lists}
\usage{
url_activities(id = NULL, friends = FALSE, club = FALSE)
}
\arguments{
\item{id}{string for id of the activity or club if \code{club = TRUE}}

\item{friends}{logical if you want friends' activities of the authenticated user}

\item{club}{logical if you want the activities of a club}
}
\value{
The set url.
}
\description{
Set the url of activities for different activity lists
}
\details{
This function concatenates appropriate strings so no authentication token is required.  This is used internally by other functions.
}
\examples{
\dontrun{
# create authentication token
# requires user created app name, id, and secret from Strava website
stoken <- httr::config(token = strava_oauth(app_name, app_client_id, 
	app_secret, cache = TRUE))

url_activities(2837007)
}
}
\concept{token}
