% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_athlete.R
\name{url_athlete}
\alias{url_athlete}
\title{Set the url of the athlete to get data}
\usage{
url_athlete(id = NULL)
}
\arguments{
\item{id}{str or integet of athlete id assigned by Strava, NULL will set the authenticated user URL}
}
\description{
Set the url of the athlete to get data using an ID
}
\details{
used by other functions
}

