% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_gear.R
\name{url_gear}
\alias{url_gear}
\title{Set the url of the equipment item to get data}
\usage{
url_gear(id)
}
\arguments{
\item{id}{string of gear id assigned by Strava}
}
\value{
A character string of the gear URL used for API requests
}
\description{
Set the url of the equipment item to get data using an ID
}
\details{
used by other functions
}
