% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_streams.R
\name{url_streams}
\alias{url_streams}
\title{Set the url for stream requests}
\usage{
url_streams(id, request = "activities", types = list("latlng"))
}
\arguments{
\item{id}{numeric for id of the request}

\item{request}{chr string defining the stream type, must be "activities", "segment_efforts", "segments"}

\item{types}{list of chr strings with any combination of "time", "latlng", "distance", "altitude", "velocity_smooth", "heartrate", "cadence", "watts", "temp", "moving", or "grade_smooth"}
}
\value{
A url string.
}
\description{
Set the url for stream requests
}
\details{
Function is used internally within \code{\link{get_streams}}. From the API documentation, 'streams' is the Strava term for the raw data associated with an activity.
}
\examples{
url_streams(123)
}
\concept{token}
