% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achievement_fun.R
\name{achievement_fun}
\alias{achievement_fun}
\title{Get recent achievements}
\usage{
achievement_fun(prsd)
}
\arguments{
\item{prsd}{parsed input list}
}
\description{
Get recent achievements, used internally in \code{\link{athl_fun}}
}
