% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_activities.R
\name{compile_activities}
\alias{compile_activities}
\title{converts a list of activities into a dataframe}
\usage{
compile_activities(list)
}
\arguments{
\item{list}{a list of all the activities you want compiling. Generally produced by \code{\link{compile_activity}}}
}
\value{
dataframe where each row is a different activity
}
\description{
converts a list of activities into a dataframe
}
\details{
each activity has a value for every column present across all activities, with NAs populating empty values
}
\examples{
 
#' stoken <- httr::config(stoken = strava_oauth(app_name, app_client_id, app_secret, cache = TRUE))
acts <- get_activity_list(stoken, 2837007)
acts_data <- compile_activities(acts)
}
\author{
Daniel Padfield
}
\concept{
posttoken
}

