% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_athlete.R
\name{get_athlete}
\alias{get_athlete}
\title{Get basic data for an athlete}
\usage{
get_athlete(stoken, id = NULL)
}
\arguments{
\item{stoken}{A \code{\link[httr]{config}} object created using the \code{\link{strava_oauth}} function}

\item{id}{string or integer of athlete}
}
\value{
Data from an API request to \url{https://strava.com/api/v3/athlete}.  Specifically, a list of athlete information including athlete name, location, followers, etc. as described here: \url{https://strava.github.io/api/v3/athlete/}.
}
\description{
Get basic athlete data for an athlete using an API request
}
\details{
Requires authentication stoken using the \code{\link{strava_oauth}} function and a user-created API on the strava website.
}
\examples{
\dontrun{
get_athlete(stoken, id = '2527465')
}
}
\concept{
token
}

