% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latlon.R
\name{get_latlon}
\alias{get_latlon}
\title{get latitude and longitude from Google polyline}
\usage{
get_latlon(x)
}
\arguments{
\item{x}{the dataframe that contains the Strava activity data}
}
\value{
dataframe of latitude and longitudes with a column for the unique identifier
}
\description{
get latitude and longitude from Google polyline
}
\examples{
\dontrun{
stoken <- httr::config(token = strava_oauth(app_name, app_client_id, app_secret, cache = TRUE))

my_acts <- get_activity_list(stoken)
acts_data <- compile_activities(my_acts)

# get lat and lon for a single activity
get_latlon(acts_data[1,])
}
}
\author{
Daniel Padfield
}
\concept{
token
}
