% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LatLon.R
\name{get_LatLon}
\alias{get_LatLon}
\title{get latitude and longitude from Google polyline}
\usage{
get_LatLon(x, .id_col)
}
\arguments{
\item{x}{the dataframe that contains the Strava activity data}

\item{.id_col}{the column that you want to be used as an identifier for the dataframe of latitude and longitude coordinates}
}
\value{
dataframe of latitude and longitudes with a column for the unique identifier
}
\description{
get latitude and longitude from Google polyline
}
\details{
used internally in \code{\link{get_all_LatLon}}
}
\author{
Daniel Padfield
}
\concept{
token
}

