% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latlon.R
\name{get_latlon}
\alias{get_latlon}
\title{get latitude and longitude from Google polyline}
\usage{
get_latlon(x)
}
\arguments{
\item{x}{the dataframe that contains the Strava activity data}
}
\value{
dataframe of latitude and longitudes with a column for the unique identifier
}
\description{
get latitude and longitude from Google polyline
}
\examples{
\dontrun{
getlatlon(act, 'upload_id')
}
}
\author{
Daniel Padfield
}
\concept{
token
}
