% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_segment.R
\name{get_segment}
\alias{get_segment}
\title{Retrieve details about a specific segment}
\usage{
get_segment(stoken, id = NULL, request = NULL)
}
\arguments{
\item{stoken}{A \code{\link[httr]{config}} object created using the \code{\link{strava_oauth}} function}

\item{id}{numeric for id of the segment}

\item{request}{chr string, must be "starred", "all_efforts", "leaderboard", "explore" or NULL for segment details}
}
\value{
Data from an API request.
}
\description{
Retreive details about a specific segment
}
\details{
Requires authentication stoken using the \code{\link{strava_oauth}} function and a user-created API on the strava website.
}
\examples{
\dontrun{
# create authentication token
# requires user created app name, id, and secret from Strava website
stoken <- httr::config(token = strava_oauth(app_name, app_client_id,
	app_secret, cache = TRUE))

get_segment(stoken, id = 123)
}
}
\concept{
token
}

