% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_fun.R
\name{monthly_fun}
\alias{monthly_fun}
\title{Get distances for last twelve months}
\usage{
monthly_fun(prsd)
}
\arguments{
\item{prsd}{parsed input list}
}
\description{
Get distances for last twelve months, used internally in \code{\link{athl_fun}}
}
\concept{notoken}
