% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava_oauth.R
\name{strava_oauth}
\alias{strava_oauth}
\title{Generata Strava API authentication token}
\usage{
strava_oauth(app_name, app_client_id, app_secret, app_scope = "public",
  cache = FALSE)
}
\arguments{
\item{app_name}{chr string for name of the app}

\item{app_client_id}{chr string for ID received when the app was registered}

\item{app_secret}{chr string for secret received when the app was registered}

\item{app_scope}{chr string for scope of authentication, Must be "public", "write", "view_private", or "view_private,write"}

\item{cache}{logical to cache the token}
}
\description{
Generate a token for the user and the desired scope. The user is sent to the strava authentication page if he/she hasn't given permission to the app yet, else, is sent to the app webpage.
}
\details{
The \code{app_name}, \code{app_client_id}, and \code{app_secret} are specific to the user and can be obtained by registering an app on the Strava API authentication page: \url{http://strava.github.io/api/v3/oauth/}.  This requires a personal Strava account.
}
\concept{token}
