% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units_fun.R
\name{units_fun}
\alias{units_fun}
\title{Get units of measurement}
\usage{
units_fun(xml_in)
}
\arguments{
\item{xml_in}{input HTMLInternalDocument from \code{\link[RCurl]{getURL}}}
}
\description{
Get units of measurement, used internally in \code{\link{athl_fun}}
}
\concept{
notoken
}

